ai-ordering_system/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
langchain==0.2.16
langchain-openai==0.1.25
python-dotenv==1.0.1
